Move all tests out of scripts folder
