+++ src/tests/unit/tests/popup/components/header-contextual-menu.test.tsx
@@ -5,15 +5,15 @@ import { ContextualMenu } from 'office-ui-fabric-react/lib/ContextualMenu';
 import * as React from 'react';
 import { It, Mock, Times } from 'typemoq';
 
-import { FeatureFlags } from '../../../../../../common/feature-flags';
-import { TelemetryEventSource } from '../../../../../../common/telemetry-events';
-import { DetailsViewPivotType } from '../../../../../../common/types/details-view-pivot-type';
-import { VisualizationType } from '../../../../../../common/types/visualization-type';
-import { PopupActionMessageCreator } from '../../../../../../popup/actions/popup-action-message-creator';
-import { HeaderContextualMenu, HeaderContextualMenuProps } from '../../../../../../popup/components/header-contextual-menu';
-import { LaunchPanelHeader } from '../../../../../../popup/components/launch-panel-header';
-import { LaunchPanelHeaderClickHandler } from '../../../../../../popup/handlers/launch-panel-header-click-handler';
-import { EventStubFactory } from '../../../../common/event-stub-factory';
+import { FeatureFlags } from '../../../../../common/feature-flags';
+import { TelemetryEventSource } from '../../../../../common/telemetry-events';
+import { DetailsViewPivotType } from '../../../../../common/types/details-view-pivot-type';
+import { VisualizationType } from '../../../../../common/types/visualization-type';
+import { PopupActionMessageCreator } from '../../../../../popup/actions/popup-action-message-creator';
+import { HeaderContextualMenu, HeaderContextualMenuProps } from '../../../../../popup/components/header-contextual-menu';
+import { LaunchPanelHeader } from '../../../../../popup/components/launch-panel-header';
+import { LaunchPanelHeaderClickHandler } from '../../../../../popup/handlers/launch-panel-header-click-handler';
+import { EventStubFactory } from '../../../common/event-stub-factory';
 
 describe('HeaderContextualMenu', () => {
     it('render', () => {
