+++ src/tests/unit/tests/popup/components/launch-pad.test.tsx
@@ -3,10 +3,10 @@
 import { shallow } from 'enzyme';
 import * as React from 'react';
 
-import { ExternalLink } from '../../../../../../common/components/external-link';
-import { toolName } from '../../../../../../content/strings/application';
-import { LaunchPad, LaunchPadDeps, LaunchPadProps, LaunchPadRowConfiguration } from '../../../../../../popup/components/launch-pad';
-import { LaunchPadItemRow } from '../../../../../../popup/components/launch-pad-item-row';
+import { ExternalLink } from '../../../../../common/components/external-link';
+import { toolName } from '../../../../../content/strings/application';
+import { LaunchPad, LaunchPadDeps, LaunchPadProps, LaunchPadRowConfiguration } from '../../../../../popup/components/launch-pad';
+import { LaunchPadItemRow } from '../../../../../popup/components/launch-pad-item-row';
 
 const AXE_CORE_VERSION = 'axe.core.version';
