+++ src/tests/unit/tests/popup/main-renderer.test.tsx
@@ -4,18 +4,18 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { IMock, It, Mock } from 'typemoq';
 
-import { BrowserAdapter, ChromeAdapter } from '../../../../../background/browser-adapter';
-import { Theme } from '../../../../../common/components/theme';
-import { DropdownClickHandler } from '../../../../../common/dropdown-click-handler';
-import { title } from '../../../../../content/strings/application';
-import { DiagnosticViewToggleFactory } from '../../../../../popup/components/diagnostic-view-toggle-factory';
-import { PopupViewWithStoreSubscription } from '../../../../../popup/components/popup-view';
-import { DiagnosticViewClickHandler } from '../../../../../popup/handlers/diagnostic-view-toggle-click-handler';
-import { LaunchPanelHeaderClickHandler } from '../../../../../popup/handlers/launch-panel-header-click-handler';
-import { PopupViewControllerHandler } from '../../../../../popup/handlers/popup-view-controller-handler';
-import { LaunchPadRowConfigurationFactory } from '../../../../../popup/launch-pad-row-configuration-factory';
-import { MainRenderer, MainRendererDeps } from '../../../../../popup/main-renderer';
-import { SupportLinkHandler } from '../../../../../popup/support-link-handler';
+import { BrowserAdapter, ChromeAdapter } from '../../../../background/browser-adapter';
+import { Theme } from '../../../../common/components/theme';
+import { DropdownClickHandler } from '../../../../common/dropdown-click-handler';
+import { title } from '../../../../content/strings/application';
+import { DiagnosticViewToggleFactory } from '../../../../popup/components/diagnostic-view-toggle-factory';
+import { PopupViewWithStoreSubscription } from '../../../../popup/components/popup-view';
+import { DiagnosticViewClickHandler } from '../../../../popup/handlers/diagnostic-view-toggle-click-handler';
+import { LaunchPanelHeaderClickHandler } from '../../../../popup/handlers/launch-panel-header-click-handler';
+import { PopupViewControllerHandler } from '../../../../popup/handlers/popup-view-controller-handler';
+import { LaunchPadRowConfigurationFactory } from '../../../../popup/launch-pad-row-configuration-factory';
+import { MainRenderer, MainRendererDeps } from '../../../../popup/main-renderer';
+import { SupportLinkHandler } from '../../../../popup/support-link-handler';
 
 describe('MainRenderer', () => {
     const expectedTitle = title;
